Added assimp and converted material loading to assimp
