+++ include/renderer.hpp
@@ -12,7 +12,7 @@ struct Buffer;
 struct Image;
 class VulkanContext;
 class SwapChain;
-class GLTFLoader;
+class ModelLoader;
 class BottomLevelAccelerationStructure;
 class TopLevelAccelerationStructure;
 class BindlessResources;
@@ -58,7 +58,7 @@ private:
 
     uint32_t _currentResourcesFrame = 0;
 
-    std::unique_ptr<GLTFLoader> _gltfLoader;
+    std::unique_ptr<ModelLoader> _modelLoader;
     std::shared_ptr<BindlessResources> _bindlessResources;
 
     std::vector<BottomLevelAccelerationStructure> _blases {};
